Refactor middleware logging to a separate file
