+++ apps/nextjs/src/middleware.ts
@@ -1,4 +1,3 @@
-import * as Sentry from "@sentry/nextjs";
 import { NextMiddlewareResult } from "next/dist/server/web/types";
 import {
   NextFetchEvent,
@@ -7,6 +6,7 @@ import {
   NextResponse,
 } from "next/server";
 
+import { logError } from "./lib/errors/server/middlewareErrorLogging";
 import { sentryCleanup } from "./lib/sentry/sentryCleanup";
 import { authMiddleware } from "./middlewares/auth.middleware";
 import { addCspHeaders } from "./middlewares/csp";
@@ -15,74 +15,26 @@ export type ErrorWithPotentialCause = Error & {
   cause?: unknown;
 };
 
-export async function handleError(
-  error: unknown,
-  request: NextRequest,
-  event: NextFetchEvent,
-): Promise<Response> {
-  const requestInfo = {
-    url: request.url,
-    method: request.method,
-    headers: Object.fromEntries(request.headers),
-    cookies: Object.fromEntries(request.cookies),
-    geo: request.geo,
-    ip: request.ip,
-    nextUrl: {
-      pathname: request.nextUrl.pathname,
-      search: request.nextUrl.search,
-    },
-  };
-
-  const eventInfo = {
-    sourcePage: event.sourcePage,
-  };
-
-  let bodyText = "";
-  try {
-    const clonedRequest = request.clone();
-    bodyText = await clonedRequest.text();
-  } catch (bodyError) {
-    console.error("Failed to read request body", bodyError);
-    bodyText = "Failed to read body";
-  }
-
-  // Cast error to our custom type
+function determineErrorResponse(error: unknown): Response {
   const errorWithCause = error as ErrorWithPotentialCause;
 
   if (
     error instanceof SyntaxError ||
     errorWithCause.cause instanceof SyntaxError
   ) {
-    const syntaxError =
-      error instanceof SyntaxError
-        ? error
-        : (errorWithCause.cause as SyntaxError);
-
-    console.warn({
-      event: "middleware.syntaxError",
-      error: syntaxError.message,
-      url: requestInfo.url,
-      request: requestInfo,
-    });
-
     return NextResponse.json({ error: "Bad Request" }, { status: 400 });
   }
 
-  const wrappedError =
-    error instanceof Error
-      ? error
-      : new Error("Error in nextMiddleware", { cause: error });
-
-  Sentry.captureException(wrappedError, {
-    extra: {
-      requestInfo,
-      eventInfo,
-      bodyText,
-      errorType: "OtherError",
-    },
-  });
-
   return NextResponse.json({ error: "Internal Server Error" }, { status: 500 });
+}
+
+export async function handleError(
+  error: unknown,
+  request: NextRequest,
+  event: NextFetchEvent,
+): Promise<Response> {
+  await logError(error, request, event);
+  return determineErrorResponse(error);
 }
 
 const nextMiddleware: NextMiddleware = async (request, event) => {
